refactor(hooks): общая логика для отрисовки страниц вынесена в хук
